Reorganize validation summarizer and problem limit
- Hide max error count and default it.
- Make summarizer class internal to exception which uses it.
