+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/ValidateStep.java
@@ -218,8 +218,7 @@ public class ValidateStep<C extends BeforeExecutionContext, R extends Result> im
                 .map(TypeValidationProblemRenderer::renderMinimalInformationAbout)
                 .collect(ImmutableSet.toImmutableSet());
         throw WorkValidationException.forProblems(uniqueErrors)
-                .limitTo(WorkValidationException.MAX_ERR_COUNT)
-                .withSummary(new ValidationErrorSummarizer(work, validationContext))
+                .withSummary(new WorkValidationException.ValidationErrorSummarizer(work, validationContext))
                 .get();
     }
